agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "opint LangGraph agent sidecar"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "pymongo>=4.9,<4.16",
    "langgraph>=1.1.10",
    "langchain-core>=1.3.2",
    "langgraph-checkpoint-mongodb>=0.3.1",
    "langchain-openai>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
